python/requirements.txt
python-socketio[client]
python-dotenv